updated project

+++ login/views.py
@@ -8,6 +8,8 @@ from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.decorators import login_required
 from .models import Product, OrderProduct, Cart, ProductPicture
 from django.core.files.storage import FileSystemStorage
+from django.core.paginator import Paginator
+from django.core.exceptions import ObjectDoesNotExist
 
 
 # Create your views here.
@@ -45,9 +47,13 @@ def productlist(request):
     elif request.GET.get('search') is not None:
         term = request.GET['search']
         product_list = Product.objects.filter(name__contains=term)
+    
+    paginator = Paginator(product_list,5)
+    page_number = request.GET.get('page')
+    page_obj = paginator.get_page(page_number)
 
     context = {
-        'products': product_list,
+        'page_obj': page_obj,
         'type': sort,
     }
 
@@ -55,13 +61,12 @@ def productlist(request):
 
 
 def post_product(request):
-    print(stripe)
-
     if request.method == "POST":
 
         if request.user.is_authenticated:
             product_name = request.POST['name']
             product_price = request.POST['price']
+            product_category = request.POST['category']
 
             if not product_name or not product_price:
                 messages.warning(request, "please enter product and price")
@@ -71,11 +76,17 @@ def post_product(request):
             files = request.FILES.getlist('product-pic')
             if files:
                 product = Product.objects.create(
-                    name=product_name, price=product_price, user=request.user)
+                    name=product_name, price=product_price, user=request.user, category=product_category)
                 for image in files:
                     product_picture = ProductPicture.objects.create(
                         picture=image, product=product)
                 return redirect('post_product')
+            else:
+                product = Product.objects.create(
+                    name=product_name, price=product_price, user=request.user, category=product_category)
+                return redirect('post_product')
+
+                
 
         else:
             messages.warning(request, "please login to post products")
@@ -86,15 +97,19 @@ def post_product(request):
 
 @login_required(login_url='/login/')
 def cart(request):
-    cart, created = Cart.objects.get_or_create(user=request.user)
+    cart, created = Cart.objects.get_or_create(user=request.user, ordered=False)
     order_product_list = cart.products.order_by('-product__price')
-    total = 0
+    subtotal = 0
     for order_product in order_product_list:
-        total += order_product.product.price * order_product.quantity
+        subtotal += order_product.product.price * order_product.quantity
+    tax = subtotal * 0.13
+    total = subtotal + tax
 
     context = {
         'products': order_product_list,
-        'total': total,
+        'subtotal': subtotal,
+        'tax' : tax,
+        'total' : total
     }
 
     return render(request, 'login/cart.html', context)
@@ -103,11 +118,11 @@ def cart(request):
 def add_to_cart(request, product_id):
     product = get_object_or_404(Product, pk=product_id)
     order_product, created = OrderProduct.objects.get_or_create(
-        product=product, user=request.user)
-    cart_qs = Cart.objects.filter(user=request.user)
+        product=product, user=request.user, ordered=False)
+    cart_qs = Cart.objects.filter(user=request.user,ordered=False)
     if cart_qs.exists():
         cart = cart_qs[0]
-        if cart.products.filter(product__name=product.name).exists():
+        if not created:
             order_product.quantity += 1
             order_product.save()
         else:
@@ -127,16 +142,19 @@ def delete_product(request, product_id):
 
 def delete_from_cart(request, product_name):
     product = get_object_or_404(Product, name=product_name)
-    order_product = OrderProduct.objects.get(
-        product=product, user=request.user)
-    cart = Cart.objects.get(user=request.user)
-    if order_product.quantity > 1:
-        order_product.quantity -= 1
-        order_product.save()
-        return redirect("cart")
+    order_product_qs = OrderProduct.objects.filter(
+        product=product, user=request.user, ordered=False)
+    if order_product_qs.exists():
+        order_product = order_product_qs[0]
+        if order_product.quantity > 1:
+            order_product.quantity -= 1
+            order_product.save()
+            return redirect("cart")
+        else:
+            order_product.delete()
+            return redirect("cart")
     else:
-        order_product.delete()
-        return redirect("cart")
+        return redirect("products")
 
 
 def register(request):
@@ -172,25 +190,95 @@ def productdetail(request, product_id):
             'rest_picture': rest_picture,
             'count': count
         }
-        return render(request, 'login/product_detail.html', context)
+        return render(request, 'login/product_detail-new.html', context)
     else:
-        return render(request, 'login/product_detail.html', {'product': product})
+        return render(request, 'login/product_detail-new.html', {'product': product})
 
 
 def checkout(request):
+    try:
+        cart = Cart.objects.get(user=request.user,ordered=False)
+        total_bill = cart.total_price()
+        if total_bill == 0:
+            messages.warning(request, "There is nothing in your cart!")
+            return redirect('products') 
+
+        try:
+            products = cart.products.all()
+            context = { 
+                'products' : products, 
+                'total' : total_bill
+                }
+            return render(request,'login/checkout.html',context)
+
+        except:
+            messages.warning(request, "There is nothing in your cart!")
+            return redirect('products')
+    except ObjectDoesNotExist:
+        messages.warning(request, "There is nothing in your cart!")
+        return redirect('products')
+
     return render(request, 'login/checkout.html', {})
 
 
-def payment(request):
-    publishableKey = settings.STRIPE_PUBLISHABLE_KEY
+def make_payment(request):
     if request.method == "POST":
         token = request.POST['stripeToken']
-        charge = stripe.Charge.create(
-            amount=999,
+
+        cart = Cart.objects.get(user=request.user, ordered=False)
+        price = cart.total_price() * 100
+
+        try:
+            # Use Stripe's library to make requests...
+            charge = stripe.Charge.create(
+            amount=price,
             currency='cad',
             description='Example charge',
             source=token,
-        )
-        
-        return render(request, 'login/payment.html', { 'publishable_key' : publishableKey})
-    return render(request, 'login/payment.html', { 'publishable_key' : publishableKey})
+            )
+        except stripe.error.CardError as e:
+            # Since it's a decline, stripe.error.CardError will be caught
+
+            print('Status is: %s' % e.http_status)
+            print('Type is: %s' % e.error.type)
+            print('Code is: %s' % e.error.code)
+            # param is '' in this case
+            print('Param is: %s' % e.error.param)
+            print('Message is: %s' % e.error.message)
+            messages.warning(request, e.error.message)
+            return redirect('products')
+        except stripe.error.RateLimitError as e:
+            # Too many requests made to the API too quickly
+            pass
+        except stripe.error.InvalidRequestError as e:
+            # Invalid parameters were supplied to Stripe's API
+            pass
+        except stripe.error.AuthenticationError as e:
+            # Authentication with Stripe's API failed
+            # (maybe you changed API keys recently)
+            pass
+        except stripe.error.APIConnectionError as e:
+            # Network communication with Stripe failed
+            pass
+        except stripe.error.StripeError as e:
+            # Display a very generic error to the user, and maybe send
+            # yourself an email
+            pass
+        except Exception as e:
+            # Something else happened, completely unrelated to Stripe
+            pass
+
+        cart.ordered = True
+        cart.save()
+        for product in cart.products.all():
+            product.ordered = True
+            product.save()
+
+        return render(request, 'login/payment.html')
+    return render(request, 'login/payment.html')
+
+
+def stripe_payment(request):
+    publishableKey = settings.STRIPE_PUBLISHABLE_KEY
+
+    return render(request, 'login/stripe-payment.html', { 'publishable_key' : publishableKey})
